fix: use ttl check with date util
